requirements.txt
apache-airflow-providers-postgres==5.2.1
psycopg2-binary>=2.8.0
